requirements.txt
flask
pandas
numpy
scikit-learn
sentence-transformers
waitress
psycopg2-binary
python-dotenv
